refactor: add feature enginerring code to a different file
